+++ test/test_cpufuncs.cpp
@@ -14,7 +14,7 @@ const float kSatFilterCoeffs[] = {1.0f, 1.0f};
 const float kArbitratyFilterCoeffs[] = {1.723f, 0.546f};
 
 TEST(CPU_funcs, naive_sat) {
-  const int n_rows = 5, n_cols = 5;
+  const size_t n_rows = 5, n_cols = 5;
   const int input_data[] = {
       1, 0, 0, 1, 1, 2, 1, 1, 1, 2, 1, 2, 1, 2, 0, 1, 1, 0, 1, 1, 0, 1, 2, 1, 2,
   };
@@ -36,7 +36,7 @@ TEST(CPU_funcs, naive_sat) {
 }
 
 TEST(CPU_funcs, filtery_sat) {
-  const int n_rows = 5, n_cols = 5;
+  const size_t n_rows = 5, n_cols = 5;
   const int input_data[] = {
       1, 0, 0, 1, 1, 2, 1, 1, 1, 2, 1, 2, 1, 2, 0, 1, 1, 0, 1, 1, 0, 1, 2, 1, 2,
   };
@@ -56,7 +56,7 @@ TEST(CPU_funcs, filtery_sat) {
 }
 
 TEST(CPU_funcs, arbitrary_coeffs_filter) {
-  const int n_rows = 5, n_cols = 5;
+  const size_t n_rows = 5, n_cols = 5;
   const int input_data[] = {
       1, 0, 0, 1, 1, 2, 1, 1, 1, 2, 1, 2, 1, 2, 0, 1, 1, 0, 1, 1, 0, 1, 2, 1, 2,
   };
